+++ packages/core/src/index.ts
@@ -84,6 +84,7 @@ export * from "./expression_belief_state.js";
 export * from "./latent_cross_influence.js";
 export * from "./routing_adaptation.js";
 export * from "./memory_rotation.js";
+export * from "./memory_forgetting.js";
 export * from "./persona_write_lock.js";
 export * from "./memory_store_driver.js";
 export * from "./persona_lint.js";